requirements.txt
opencv_python_headless==4.7.0.72
pandas==1.5.3
Pillow==9.5.0
pytesseract==0.3.10
python-dotenv==1.0.0
streamlit==1.21.0
